backend/requirements.txt
crewai==1.14.4
fastapi==0.136.1
uvicorn[standard]==0.46.0
anthropic==0.101.0
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-settings==2.10.1
litellm==1.83.14
openai==2.36.0
instructor==1.15.1
numpy==2.4.4
a2a-sdk==1.0.3
google-cloud-secret-manager==2.22.0
